feat: show more shows on showlist
