--- src/components/ShowPreview/styles.scss
@@ -1,43 +1,0 @@
-@import 'variables.scss';
-
-.container {
-  background-color: #FFF;
-  border-radius: 1px;
-  border-top: 3px solid $color-radiorevolt-orange;
-  box-shadow: -1px 0 2px 0 rgba(0, 0, 0, 0.12), 1px 0 2px 0 rgba(0, 0, 0, 0.12), 0 1px 1px 0 rgba(0, 0, 0, 0.24);
-  margin: 1rem;
-  padding: 1rem;
-  width: 20rem;
-}
-
-.image {
-  width: 100%;
-}
-
-.nameLink {
-  color: inherit;
-  text-decoration: none;
-}
-
-.nameLink:hover {
-  text-decoration: underline;
-}
-
-.name {
-  text-align: center;
-}
-
-.logoImage {
-  display: flex;
-  flex-direction: column;
-  height: 315px;
-  justify-content: center;
-}
-
-.category {
-  border-radius: 0.3em;
-  margin-left: -1.5em;
-  margin-top: -0.5em;
-  padding: 0.1em 0.5em;
-  position: absolute;
-}
