--- src/components/ShowPreviewList/styles.css
@@ -1,38 +1,0 @@
-.activeShows {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  margin-bottom: 2rem;
-}
-
-.archivedShowsButton {
-  font-size: 24px;
-  border-width: 0;
-}
-
-.arrowDown {
-  width: 20px;
-  transition: 100ms linear all;
-  transform: rotate(90deg);
-}
-
-.arrowLeft {
-  width: 20px;
-  transition: 100ms linear all;
-}
-
-.archivedShowsButton:hover {
-  cursor: pointer;
-}
-
-.archivedShowsVisible {
-  display: flex;
-  flex-wrap: wrap;
-  height: auto;
-  justify-content: space-around;
-}
-
-.archivedShowsHidden {
-  display: none;
-  overflow: hidden;
-}
