+++ src/components/Shows/ShowPreview/styles.scss
@@ -1,0 +1,48 @@
+@import 'variables.scss';
+
+$max-width: 340px;
+$padding: 14px;
+$image-height: $max-width - 2 * $padding;
+
+.container {
+  display: flex;
+  flex-direction: column;
+  background-color: #FFF;
+  border-top: 3px solid $color-radiorevolt-orange;
+  box-shadow: -1px 0 2px 0 rgba(0, 0, 0, 0.12), 1px 0 2px 0 rgba(0, 0, 0, 0.12), 0 1px 2px 0 rgba(0, 0, 0, 0.12);
+  padding: $padding;
+  height: 100%;
+  max-width: $max-width;
+}
+
+.imageWrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: $image-height;
+}
+
+.image {
+  width: 100%;
+  height: auto;
+  max-height: $image-height;
+}
+
+.title {
+  color: inherit;
+  text-decoration: none;
+  text-align: center;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.category {
+  border-radius: 0.3em;
+  margin-left: -1.5em;
+  margin-top: -0.5em;
+  padding: 0.1em 0.5em;
+  position: absolute;
+}
